Added a Langevin example notebook

+++ diffrax/solver/align.py
@@ -128,7 +128,7 @@ def _eval_taylor(h, tay_cfs):
     # Multiplies the pre-computed Taylor coefficients by powers of h.
     # jax.debug.print("eval taylor for h = {h}", h=h)
     dtype = jnp.dtype(tay_cfs["a1"])
-    h_powers = jnp.power(h, jnp.arange(0, 6, dtype=dtype))
+    h_powers = jnp.power(h, jnp.arange(0, 6)).astype(dtype)
     return jtu.tree_map(
         lambda tay_leaf: jnp.tensordot(tay_leaf, h_powers, axes=1), tay_cfs
     )
